build_helper::ci
Module gha

Structs

Group

A guard that closes the current github actions log group on drop.

Statics

ACTIVE_GROUPS 🔒

Functions

end_group 🔒

group

All github actions log messages from this call to the Drop of the return value will be grouped and hidden by default in logs. Note that since github actions doesn’t support group nesting, any active group will be first finished when a subgroup is started, and then re-started when the subgroup finishes.

is_in_gha 🔒

start_group 🔒