Crate build_helper

Types and functions shared across tools in this workspace.

Modules

ci

drop_bomb

This module implements “drop bombs” intended for use by command wrappers to ensure that the constructed commands are eventually executed. This is exactly like rustc_errors::Diag where we force every Diag to be consumed or we emit a bug, but we panic instead.

fs

Misc filesystem related helpers for use by bootstrap and tools.

git

metrics

npm

stage0_parser

targets

util

Macros

exit

Invokes build_helper::util::detail_exit with cfg!(test)

Constants

LLVM_PGO_CRATES

The default set of crates for opt-dist to collect LLVM profiles.

RUSTC_PGO_CRATES

The default set of crates for opt-dist to collect rustc profiles.